# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Delivery API",
        "description": "API for Food Delivery Application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "responses": {
                    "200": {
                        "description": "List of orders"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "restaurant_id": {
                                        "type": "integer"
                                    },
                                    "total_price": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "payment_status": {
                                        "type": "string"
                                    },
                                    "delivery_address": {
                                        "type": "string"
                                    },
                                    "placed_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "completed_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a specific order",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "restaurant_id": {
                                        "type": "integer"
                                    },
                                    "total_price": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "payment_status": {
                                        "type": "string"
                                    },
                                    "delivery_address": {
                                        "type": "string"
                                    },
                                    "placed_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "completed_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "operationId": "303a7b797887b66c7790c4c187b54790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/order-items": {
            "get": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Get all order items",
                "operationId": "6a6b106e104d558bfc813be8f4adf2f4",
                "responses": {
                    "200": {
                        "description": "List of order items"
                    }
                }
            },
            "post": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Create a new order item",
                "operationId": "3390f5f422c701c6e520cc66e9d18037",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "menu_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order item created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/order-items/{id}": {
            "get": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Get a specific order item",
                "operationId": "f75b95ab4f85b8772c2981a0d6e449bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order item details"
                    },
                    "404": {
                        "description": "Order item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Update an order item",
                "operationId": "c1778023c9a3ecbc744efb765d68415f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "menu_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order item updated successfully"
                    },
                    "404": {
                        "description": "Order item not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Delete an order item",
                "operationId": "87d21d8e8c19f5d96b514a9aa273dede",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order item deleted successfully"
                    },
                    "404": {
                        "description": "Order item not found"
                    }
                }
            }
        },
...
    },
...
}